This is an Android App made with android studio.
This app is the "Apps" section of question 4. I wasn't able to add new features, but if I had the time I would a push notification feature that allowed users to subscribe for alerts every time an article is posted or just to alerts about specific topics or other articles. Those alerts would be in the form of push notifications.